requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pycryptodome==3.18.0
kyber-py
dilithium-py
pytest
requests